Authentication

+++ internal/usecases/playlist/playlist.go
@@ -1,6 +1,9 @@
 package playlist
 
-import "github.com/mp-hl-2021/muzio/internal/domain/playlist"
+import (
+	"github.com/mp-hl-2021/muzio/internal/domain"
+	"github.com/mp-hl-2021/muzio/internal/domain/playlist"
+)
 
 type Playlist struct {
 	Name    string
@@ -10,8 +13,8 @@ type Playlist struct {
 type Interface interface {
 	CreatePlaylist(owner, name string, content []string) (string, error)
 	GetPlaylistById(id string) (Playlist, error)
-	UpdatePlayList(/* owner, */ id, name string, content []string) error
-	DeletePlayList(/* owner, */ id string) error
+	UpdatePlayList(owner, id, name string, content []string) error
+	DeletePlayList(owner, id string) error
 }
 
 type UseCases struct {
@@ -19,7 +22,6 @@ type UseCases struct {
 }
 
 func (u *UseCases) CreatePlaylist(owner, name string, content []string) (string, error) {
-	// TODO: Auth
 	p, err := u.PlaylistStorage.CreatePlaylist(owner, name, content)
 	if err != nil {
 		return "", err
@@ -35,12 +37,24 @@ func (u *UseCases) GetPlaylistById(id string) (Playlist, error) {
 	return Playlist{Name: p.Name, Content: p.Content}, nil
 }
 
-func (u *UseCases) UpdatePlayList(/* owner, */ id, name string, content []string) error {
-	// TODO: Auth
+func (u *UseCases) UpdatePlayList(owner, id, name string, content []string) error {
+	p, err := u.PlaylistStorage.GetPlaylistById(id)
+	if err != nil {
+		return err
+	}
+	if p.Owner != owner {
+		return domain.ErrForbidden
+	}
 	return u.PlaylistStorage.UpdatePlaylist(id, name, content)
 }
 
-func (u *UseCases) DeletePlayList(/* owner, */ id string) error {
-	// TODO: Auth
+func (u *UseCases) DeletePlayList(owner, id string) error {
+	p, err := u.PlaylistStorage.GetPlaylistById(id)
+	if err != nil {
+		return err
+	}
+	if p.Owner != owner {
+		return domain.ErrForbidden
+	}
 	return u.PlaylistStorage.DeletePlaylist(id)
 }
